Module mcap::read

Read MCAP files

MCAPs are read from a byte slice instead of a Read trait object. This helps us avoid unnecessary copies, since Schemas and Messages can refer directly to their data.

Consider memory-mapping the file - the OS will load (and cache!) it on-demand, without any further system calls.

Structs

ChunkFlattener

Like LinearReader, but unpacks chunks’ records into its stream

ChunkReader

Streams records out of a Chunk, decompressing as needed.

LinearReader

Scans a mapped MCAP file from start to end, returning each record.

MessageStream

Like RawMessageStream, but constructs a Message (complete with its Channel) from the raw header and data.

RawMessage

RawMessageStream

Reads all messages from the MCAP file—in the order they were written—and perform needed validation (CRCs, etc.) as we go.

Summary

Indexes of an MCAP file parsed from its (optional) summary section

Enums

Options

Nonstandard reading options, e.g., to be more lenient when trying to recover incomplete/damaged files.

Functions

attachment

Read the attachment with the given index.

footer

Read the MCAP footer.

metadata

Read the metadata with the given index.